qrcode/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
